Move the AbstractStructureTaggerProvider down to EditorFeatures
This consolidates the implementation between .Cocoa and .Wpf.
