+++ src/EditorFeatures/Core/Implementation/Structure/InvalidOutliningRegionException.cs
@@ -2,8 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-#nullable disable
-
 using System;
 using Microsoft.CodeAnalysis.Structure;
 using Microsoft.VisualStudio.Text;
