+++ src/EditorFeatures/Core/Implementation/Structure/OutliningCommandHandler.cs
